move baselineTracking and baselines and update the others accordingly.
