+++ ProcessingAlgorithms/BoundingAreaOfInterest/baselines.py
