+++ peak_follower.py
@@ -9,7 +9,7 @@
 
 import numpy as np
 from follower import Follower
-from baselines import baselines_by_squash as bline
+from ProcessingAlgorithms.BoundingAreaOfInterest.baselines import baselines_by_squash as bline
 from ProcessingAlgorithms.Fitting.moving_average import moving_average
